backend/requirements.txt
# ── Core ──────────────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# ── Database ──────────────────────────────────────────────
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# ── CORS / Headers ────────────────────────────────────────
django-cors-headers==4.3.1

# ── Environment ───────────────────────────────────────────
python-decouple==3.8

# ── Image handling ────────────────────────────────────────
Pillow==10.3.0

# ── Filtering ─────────────────────────────────────────────
django-filter==24.2

# ── Email ─────────────────────────────────────────────────
django-anymail==10.3

# ── Storage (S3 — optional, set USE_S3=True) ──────────────
django-storages==1.14.3
boto3==1.34.69

# ── API docs ──────────────────────────────────────────────
drf-yasg==1.21.7

# ── Production server ─────────────────────────────────────
whitenoise==6.6.0
gunicorn==22.0.0

# ── Dev utilities ─────────────────────────────────────────
django-extensions==3.2.3
